requirements.txt
Django>=4.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
pillow>=10.0.0
celery
redis
django-celery-beat